Invoice form and results added
Components used to add the details of an invoice and to search invoices respectively. The invoice form component uses: invoice detail form, invoice detail results, payment form, payment results and the discount form components.

The component is prepared to receive an input (existent invoice) to modify it.

Dates have not been implemented, yet.