View invoice component created
Component made to view an invoice with its information.

invoice-detail-results: Can delete lines when updating an invoice and hides the delete button when we are viewing the invoice.
payment-results: Hides the overturn button when we are viewing the invoice.
invoice-service: typos.
app-routing-module: Routing for the component added.
app.module: Component added