Search invoices uses query parameters (dates)
Search invoices now uses query parameters (start and an end date).
Invoice results now receives a valid date range as a default value.
Date range component adjusted to receive default values.
Date service adds functions to validate dates and to set the time on a date to the start or the end of the day.