Order form fixes
Fixed an issue where the search provider didn't display a dialog that was supposed to.
Fixed an issue where the payment component wouldn't receive any changes made on payments.
Fixed an error where it wasn't possible to put/remove dates when updating an order because the date range components received an ISO string instead of a date object.
Corrected a typo in an output variable in the payment results component. It no longer how much was owed after and before the payment.
Corrected the same typo in the payment form.
Corrected the same typo in the invoice form.